refactor: move Schlage, Akuvox, and Virtual provider tests into dedicated modules
Mirror the directory structure already used by matter/ and zwave_js/ providers.
Each provider now has its own __init__.py, conftest.py (with fixtures), and
test file under tests/providers/<provider>/.
