+++ tests/providers/akuvox/__init__.py
@@ -1,0 +1,1 @@
+"""Akuvox provider tests."""
