+++ tests/providers/schlage/__init__.py
@@ -1,0 +1,1 @@
+"""Schlage provider tests."""
