+++ tests/providers/virtual/__init__.py
@@ -1,0 +1,1 @@
+"""Virtual provider tests."""
